Fixed lighting for moon, cloud, multiple point lights.
Took the emissive calculation out of the diffuse and specular function in the shader and use it to add the Ke values to the final color, same as the ambient light calculation. Fixed not being able to add more lights to the scene. I am unsure why the changes in the shaders work/didn't work before, but it works now so woohoo!