Fixed multiple warnings. Added Moon texture.
Multiple warnings and messages were fixed. Moon texture was added, however, there is still an issue where the moon does not show texture 2 (the moon), but rather texture 0 (win). I have checked and the moon texture does work, it's just that the moon sphere is not using the correct one for some reason.